% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{add_path}
\alias{add_path}
\title{Modify parts of the ramp path by name.}
\usage{
add_path(
  rpath,
  project = NULL,
  stage = NULL,
  dataset = NULL,
  version = NULL,
  file = NULL
)
}
\arguments{
\item{rpath}{The original ramp path.}

\item{project}{The project name as a string.}

\item{stage}{The stage name.}

\item{dataset}{The dataset name.}

\item{version}{The version name.}

\item{file}{The filename, which can be a relative path to a file within the
directory.}
}
\value{
A ramp path with the specified parts modified.
}
\description{
Modify parts of the ramp path by name.
}
