% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{as.lfn}
\alias{as.lfn}
\title{Return a path as a logical filename, which is a relative path.}
\usage{
as.lfn(rpath)
}
\arguments{
\item{rpath}{A RAMP path.}
}
\value{
A string with forward-slashes separating parts of the filename.
}
\description{
Return a path as a logical filename, which is a relative path.
}
