% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{cached_config}
\alias{cached_config}
\title{Get the configuration file and store as package global.}
\usage{
cached_config(key, builder)
}
\description{
Get the configuration file and store as package global.
}
