% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{data_configuration}
\alias{data_configuration}
\title{Reads configuration file on where to download data.}
\usage{
data_configuration(section = "Default")
}
\arguments{
\item{section}{Which section of the initialization file to read.
This defaults to \code{Default}. This parameter exists in order
to create a \code{Test} section to use for non-destructive testing.}
}
\value{
A list of configuration parameters.
}
\description{
If you want to make the config file, then create
a file in one of three places:
}
\details{
- \code{${XDG_CONFIG_HOME}/RAMP/data.ini}
- \code{$HOME/.config/RAMP/data.ini}
- \code{$HOME/.ramp.ini}

Put the following in it.
}
\examples{
\dontrun{
[Default]
SCPHOST = computer-name.ihme.uw.edu
SCPHOSTBASE = /path/to/data/directory
LOCALDATA = /home/username/data
}

}
