% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\docType{data}
\name{.rampdata.config}
\alias{.rampdata.config}
\title{This variable contains configuration for the file locations.}
\format{
An object of class \code{list} of length 0.
}
\usage{
.rampdata.config
}
\description{
This variable contains configuration for the file locations.
}
\keyword{datasets}
