% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{download_worldpop}
\alias{download_worldpop}
\title{Retrieve worldpop data.}
\usage{
download_worldpop(local_directory = "inst/extdata", overwrite = FALSE)
}
\arguments{
\item{overwrite}{Whether to overwrite an existing file by the same name.}

\item{local}{directory Where to put that file on the local machine.}
}
\description{
Worldpop returns several GeoTIFFs in WGS 84.
GNQ = Equatorial Guinea
10 or 15 is 2010 or 2015 data.
adjv2 or v2 is whether it was adjusted to match WHO.
So use GNQ15v2.tif.
}
