% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{ensure_on_server}
\alias{ensure_on_server}
\title{Ensure a set of local files are on the server.}
\usage{
ensure_on_server(ramp_identifiers)
}
\arguments{
\item{ramp_identifiers}{A list of RAMP IDs.}
}
\description{
Ensure a set of local files are on the server.
}
