% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{ensure_present}
\alias{ensure_present}
\title{Ensure files are available locally.}
\usage{
ensure_present(ramp_identifiers)
}
\arguments{
\item{ramp_identifiers}{A list of RAMP IDs.}
}
\description{
Ensure files are available locally.
}
