% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{get_from_ihme}
\alias{get_from_ihme}
\title{Use scp to retrieve data.}
\usage{
get_from_ihme(session, filename, data_configuration = NULL)
}
\arguments{
\item{session}{An ssh session.}

\item{filename}{The path of the file within the repository.}

\item{local_directory}{Where to put that file on the local machine.}
}
\description{
Only if you have ssh credentials set up.
This is used for data that isn't yet public.
Equivalent to: ssh ihme.uw.edu:/path/to/file local_file.dat
}
\details{
You have to call `data_configuration()` before you call this.
}
