% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{initialize_workflow}
\alias{initialize_workflow}
\title{Read a workflow file in order to set up versions of inputs.}
\usage{
initialize_workflow(filename)
}
\arguments{
\item{filename}{The filename with the workflow in it.
This reads a key-value file with names of versions and files.}
}
\description{
Read a workflow file in order to set up versions of inputs.
}
