% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organization.R
\name{inverse_ramp_path}
\alias{inverse_ramp_path}
\title{Turns a relative path into a ramp path.}
\usage{
inverse_ramp_path(path)
}
\arguments{
\item{path}{A path relative to the base data directory.}
}
\value{
A list that is a ramp path.
}
\description{
Turns a relative path into a ramp path.
}
\examples{
inverse_ramp_path("users/ad") == list(stage = "working", user = "ad")
}
