% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{local_path}
\alias{local_path}
\title{The fully-rooted local path to a file.}
\usage{
local_path(file_id, config = NULL)
}
\arguments{
\item{file_id}{A list identifying the file.}

\item{config}{A configuration from \link{data_configuration}.
If this is NULL, then a new one is read from your dotfiles.}
}
\value{
A fully-rooted path.
}
\description{
The fully-rooted local path to a file.
}
