% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organization.R
\name{project_path}
\alias{project_path}
\title{Creates a path relative to the project root, following}
\usage{
project_path(country, project, stage)
}
\arguments{
\item{country}{If this is associated with a country. Can be NULL.}

\item{project}{The Github project or R project name, without a slash.
Can be NULL.}

\item{stage}{One of (source, user, prod). Must not be NULL.}
}
\value{
a path to use as a base path for this dataset.
}
\description{
Creates a path relative to the project root, following
}
