% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{ramp_path}
\alias{ramp_path}
\title{Make a RAMP path.}
\usage{
ramp_path(
  rooted_path = NULL,
  project = NULL,
  stage = NULL,
  dataset = NULL,
  version = NULL,
  file = NULL
)
}
\arguments{
\item{rooted_path}{Optional. Specify a path with a single string that looks like
a rooted path. This could be "/" for an empty path that you augment later.
It could be "/project/stage/dataset/version/filename".}

\item{project}{Optional project name.}

\item{stage}{Optional stage.}

\item{dataset}{Optional dataset.}

\item{version}{Optional version.}

\item{file}{Optional filename, which is relative path, including subdirectory to
a file.}
}
\value{
a ramp path.
}
\description{
Make a RAMP path.
}
\examples{
ramp_path("/uganda2020/mic/population/201015_final/population.csv")
ramp_path(project = "uganda2020", dataset = "population")
}
