% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{ramp_path_from_rooted}
\alias{ramp_path_from_rooted}
\title{Convert a single string into a ramp path.}
\usage{
ramp_path_from_rooted(rooted_path)
}
\arguments{
\item{rooted_path}{Must start with a "/". If it starts with "inputs",
then it's in the global area. Otherwise, it's in a project.}
}
\value{
a list describing a ramp path.

The order is `/project/stage/dataset/version/path`.
}
\description{
Convert a single string into a ramp path.
}
