% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rampdata.R
\docType{package}
\name{rampdata}
\alias{rampdata}
\title{rampdata: put and get data for RAMP calculations}
\description{
This package is responsible for transfer of data for
processing by RAMP projects. Data is either on the institute's
repository or on cloud storate. This transfers that data to
your machine on request, or it transfers that data from your
machine to the appropriate place on the repository.
}
