% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{read_workflow_config}
\alias{read_workflow_config}
\title{Read a workflow description file.}
\usage{
read_workflow_config(filename)
}
\arguments{
\item{filename}{The path to the config file to read.}
}
\value{
A list of config information from `configr`.
}
\description{
Read a workflow description file.
}
