% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{save_source}
\alias{save_source}
\title{Save information about a directory or data file.}
\usage{
save_source(path, properties)
}
\arguments{
\item{path}{The path to the directory or file.}

\item{properties}{A list of information about that path.}
}
\description{
If you give it a directory, it saves a file in the directory.
If you give it a filename, it saves a file next to that one.
}
\details{
The format is TOML. The file ends in .rampmd (?).
The document structure follows W3C-Prov ontology, as best we
can, which means the TOML has subheadings for things that are
agents, activities, and entities.

I would like to fill out some properties automatically, when
possible.

Some well-known properties:

\itemize{
  \item title Something you would call this data.
  \item download_date When the file was
  \item creator Person or organization that made the file.
  \item creator_email Email of person or organization that made the file.
  \item obtainer Person who got the file.
  \item obtainer_email Email of person who got the file.
  \item format Describe the data format.
  \item source_repository The git repository of code that made the data.
  \item source_version Version number of the source code.
  \item source_hash Hash of the git checkout.
  \item source_branch Branch of the code that made the data.
  \item description A free text description of what's in the file.
  \item creation_date When the file was created.
}
}
