% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{send_to_ihme}
\alias{send_to_ihme}
\title{Use scp to send data.}
\usage{
send_to_ihme(session, filename, local_directory = "inst/extdata")
}
\arguments{
\item{session}{An ssh session.}

\item{filename}{The path of the file within the repository.}

\item{local_directory}{Where to put that file on the local machine.}
}
\description{
Use scp to send data.
}
