% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R
\name{un7zip}
\alias{un7zip}
\title{Unzip a 7zip file.}
\usage{
un7zip(archive, where)
}
\arguments{
\item{archive}{Path to the file in 7z format.}

\item{where}{Directory into which to unzip the archive.
This command changes the working directory to `where`
in order to unzip.}
}
\description{
There is a Github project called archive
that would do this, too. This command is on Ubuntu but may
not be elsewhere.
}
