% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{workflow_path}
\alias{workflow_path}
\title{Get a ramp_path from a workflow configuration file.}
\usage{
workflow_path(role)
}
\arguments{
\item{role}{The key used to identify the path in the configuration file.}
}
\value{
a ramp_path made from that key.
}
\description{
Get a ramp_path from a workflow configuration file.
}
\seealso{
\code{\link{initialize_workflow}}
}
