% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{as.path}
\alias{as.path}
\title{Convert a ramp path into a simple string filename for reading or writing.}
\usage{
as.path(rpath)
}
\arguments{
\item{rpath}{A ramp path.}
}
\value{
A character string.

This applies the configuration file information to find the root of
the data directory and place the filename in that root.
}
\description{
Convert a ramp path into a simple string filename for reading or writing.
}
