% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organization.R
\name{project_path}
\alias{project_path}
\title{Creates a path relative to the project root}
\usage{
project_path(file_id)
}
\arguments{
\item{file_id}{A list with keys (stage, location, project, user, rproject).
If user is missing, the effective user is supplied.
Only stage is required.}
}
\value{
a path to use as a base path for this dataset.
}
\description{
Creates a path relative to the project root
}
\seealso{
\link{ramp_path}
}
