% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organization.R
\name{ramp_path}
\alias{ramp_path}
\title{Formats a base path for a file.}
\usage{
ramp_path(stage, location, project, user, rproject, path = NULL)
}
\arguments{
\item{stage}{One of (input, working, output). Must not be NULL.}

\item{location}{If this is associated with a location. Can be NULL.}

\item{project}{The Github project or R project name, without a slash.
Can be NULL.}

\item{user}{The username for the person's working directory.}

\item{rproject}{The Rstudio project name.}
}
\value{
a path to use as a base path for this dataset.
}
\description{
This function reflects how we want to organize files.
It returns a path relative to whatever is the
data root.
}
