% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organization.R
\name{ramp_path}
\alias{ramp_path}
\title{Formats a base path for a file.}
\usage{
ramp_path(country, project, stage)
}
\arguments{
\item{country}{If this is associated with a country. Can be NULL.}

\item{project}{The Github project or R project name, without a slash.
Can be NULL.}

\item{stage}{One of (source, user, prod). Must not be NULL.}
}
\value{
a path to use as a base path for this dataset.
}
\description{
This function reflects how we want to organize files.
It returns a path relative to whatever is the
data root.
}
