% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{send_to_ihme}
\alias{send_to_ihme}
\title{Use scp to send data.}
\usage{
send_to_ihme(session, filename, overwrite = TRUE, data_configuration = NULL)
}
\arguments{
\item{session}{An ssh session.}

\item{filename}{The path of the file within the repository.}

\item{overwrite}{Whether it is OK to overwrite the destination file.}

\item{local_directory}{Where to put that file on the local machine.}
}
\description{
Responsible for moving files within a certain set of remote
and local directory structures, as specified by the data configuration.
This will create directories and ensure that files aren't
overwritten by the copy command.
}
